Cache a saját fájljába, meg opcionálisak a cuccai.
